Add files via upload
Quora Question Pair Similarity
Quora is a online platform for asking questions and providing answers.Its head quarter is in Mountain View ,California.It was founded on june 25,2009 and made available to public on june 21,2010. So in todays generation, quora has emerged as a very huge and usefull platform which is used in different fields.
Problem Statement:
Whenever the question is asked in quora it might be possible that this question or any similar questions has been already asked and solved in quora.So in order to provide ease to users we need to identify whether the given pair of question is similar or not. By help of this any user can get the answer to their question very quickly if we are able to find whether any question is similar to his/her question or not